Fix ineffective when adding URL parameters
My check whether we were looking at container/create was too tight,
looking for the URL suffix 'container/create'. However this allows
bypassing because the URL parameters are included in the request URI and
they change the suffix. Fix this by checking if the whole URL _contains_
'container/create'. In the future we probably want to parse the URL
completely